Add Tymoji to a dropdown menu
Thought this might convenience users. Some text does go out of the boundry, but we just need to create a list of aliases for them , and use that text (also that could be used for alt text).

#Problem
It seems to think this is a type of scratchblock, so creates the tags. I dunno, I'll fix that later. I just wanted to show you this at an early stage

#This means, do not merge yet, until we work out the scratch-blocks issue